+++ server/src/tests/mod.rs
@@ -9,6 +9,7 @@ pub mod loc_empty;
 pub mod loc_version;
 pub mod oob_regiments;
 pub mod parser_skip;
+pub mod provinces;
 pub mod registry;
 pub mod scripted_loc;
 pub mod technology;